Improving the speaking-rate visual feedback.
1) Added a new parameter (sylDurRange_R) in the expt_config files, to
control the rate range for the rhythm trials independently. This was done
because of the observation of the trend toward faster speaking rate under
the R condition in the pilot subjects.

2) Added anther new option: (showRateWarn_run_R_sameType), to allow the
restriction of the warning messages to the consecutive R trials in the
main runs, in order to reduce confusion.