Changed the bounds of formant calculation
Extended the right boundary of vowel formant calculation from the end of
the vowel (in ASR phone alignment results) to the end of the trailing
phoneme).

The rationale is that in the ASR phone alignment results, the onset of the
trailing stop consonant often creeps into the vowel a little bit.
Thus ending the formant tracking right at the end of the ASR-aligned vowel
would lead to loss of critical information.